config: set config status after "post_apply" phase
Previously, the config status was set after the "apply" phase, and then
set again after the "post_apply" phase. This can lead to a situation
where the status becomes "ready" after the "apply" phase and changes to
"check_errors" or "check_warnings" after the "post_apply" phase.
Because of this, it was possible that config:reload() may be called from
a role or an app even if though the config loading is not finished.

Note that due to this change, the "startup_in_progress" and
"reload_in_progress" statuses are now expanded to include the
"post_apply" phase.

Also note that the status still changes before the "extras.post_apply"
stage.

Needed for https://github.com/tarantool/tarantool-ee/issues/643

NO_DOC=internal