config: introduce sharding.weight
Closes #9775

@TarantoolBot document
Title: The `sharding.weight` option

The `sharding.weight` option can be set to facilitate rebalancing. The
larger the weight of a replicaset, the more data it can store. This is a
relative value, so a replicaset with a value N times larger than a value
of the another replicaset can store N times more data. The default value
is 1.