net.box: do not start worker fiber synchronously
Due to a regression introduced in c13b3a31, the worker fiber is started
synchronously, while it should be started asynchronously, in order for the
`wait_connected = false` option of `connect` to work correctly. We already
explicitly wait from Lua for the connection to become active via
`wait_state` when `wait_connected = true`.

Closes #9489

NO_DOC=<bugfix>