box: support _origin in grant() and revoke()
This patch adds a new option '_origin' to the 'box.schema.user.grant()',
'box.schema.role.grant()', 'box.schema.user.revoke()' and
'box.schema.role.revoke()' functions. This is a private option that
allows to differentiate privileges based on their origin. If this option
is not provided an empty string is used by default.

Privileges that were granted using 'grant()' can be revoked using
'revoke()' with the same '_origin'.

The motivation for this change is to avoid actually revoking a privilege
granted by two actors if one actor revokes it.

Needed for #9643

NO_DOC=internal
NO_CHANGELOG=internal