config: introduce labels to schema
Closes #9809

@TarantoolBot document
Title: New labels in config schema

The new labels are basically maps with string keys and values,
that are merged down to instance level. It means that if a replicaset
and an instance inside it have label with same key and different
value, the actual value is the one instance provided, e.g.

```yaml
groups:
  group-001:
    replicasets:
      replicaset-001:
        labels:
          foo: 'true'
          bar: 'true'
        instances:
          instance-001:
            labels:
              baz: 'true'
              foo: 'false'
```

Results in:

```
./instance-001.iproto> require('config'):get('labels')
---
- baz: 'true'
  foo: 'false'
  bar: 'true'
...
```