console: make __autocomplete retrieval for cdata exception-safe
Currently, the code which tries to retrieve the __autocomplete metamethod
for cdata (cf644abd) is not exception safe: an exception can be thrown from
the __index metamethod. In order to make it exception-safe and not
overcomplicate the code, let's add a Lua wrapper for table indexation, so
we can call it from C using `lua_pcall`.

Closes #9828

NO_CHANGELOG=<not released>
NO_DOC=<bugfix>