box: populate port_c with type iterable
We are going to use port_c for user-defined triggers, so we need to add
an ability to store iterator in port_c. Also, this port already allows
to be dumped several times - we should keep this invariant to call
chains of triggers using only one port.

Therefore, this commit populates port_c with new type called `iterable`.
An iterable object is an object that can create iterators - an iterator
is created on each dump. We decided that such approach is more
convenient than copying an iterator in initial state on each dump.

Iterable object and created iterator have no destructor and any
invalidation logic, so author of pushed object must think about iterator
invalidation by himself. What's about destructors, user of port_c must
ensure that the object is alive as long as the port can be dumped. When
port_c is not needed anymore, the object can be deleted by caller.
Created iterator has no destructor because we think that iterator
shouldn't own any resources - its state is allocated when the iterable
object is dumped, and the release of the state lies on the side that dumped
the object. Rationale: if you dump iterable object to Lua, the
iterator is most likely to be freed only when Lua VM will run out of
memory (GC will be called then), so Tarantool resources will be occupied
for a long time. Let's try to avoid such situation.

NO_CHANGELOG=internal
NO_DOC=internal