sql: decrease number of tests in in2.test.lua
This patch reduces the number of tests in the in2.test.lua test file.
This patch also reduces the number of inserted values. This shouldn't
affect the test since it's not really an original Tarantool test, but it
will reduce the execution time of this test. Currently this test often
fails due to a timeout.

NO_DOC=test
NO_CHANGELOG=test