box: consider committed transactions as non-continuable
When we are going to start a new statement in transaction, we check if
current transaction can continue with a special helper. This helper
fails only when the transaction is aborted, but if the transaction is
committed, it actually cannot be continued as well because it is over.
So let's consider committed transactions as non-continuable.

Along the way, let's fix an error thrown when someone tries to access
database within on_rollback trigger - now it is reported as "Unknown
error", let's use "Transaction was rolled back" instead.

Closes #9186

NO_CHANGELOG=see next commit
NO_DOC=bugfix