config: rework privilege revocation
Closes #9643

@TarantoolBot document
Title: config: changes in `credentials` section

Now the privileges that were not granted by the configuration, as well
as privileges that were not granted solely by the configuration, are not
revoked on reload. Privileges that have been granted only by the config
module will still be revoked if they are removed from the `credentials`
section on reload.