config: ignore SCHEMA_NEEDS_UPGRADE error
Currently, if we try to run Tarantool 3.0 with config using old
snapshot, we may get a SCHEMA_NEEDS_UPGRADE error because granting and
revoking privileges are DDL operations. This leads to a situation where
loading Tarantool to perform an upgrade becomes quite problematic. To
avoid the issue, this patch causes 'credentials.lua' to issue a warning
instead of an error in case of the SCHEMA_NEEDS_UPGRADE error during
granting and revoking privileges.

Note that it was still possible to startup and perform the upgrade by
removing the 'credentials' section from the config or without using
config.

This is only a part of the solution, the issue will be fixed in #9849.

Part of #9849
Needed for #9643

NO_DOC=will be added later
NO_CHANGELOG=will be added later