txn_limbo: introduce limits on synchro queue
A new counter is added to `txn_limbo` structure: queue_size (in bytes)
together with a corresponding configuration setting:
`replication_synchro_queue_max_size`. The counter is increased on every
enqueued `txn_limbo_entry`, and decreased once an entry leaves the
`txn_limbo.queue`. Also, a pointer to `journal_entry` has been added to
the `txn_limbo_entry` structure, so that at the time of adding/deleting
an entry to the queue, we have access to the size of the corresponding
entry in the journal.

Actually, the limit is added to an abstract `txn_limbo`, but currently
only works for the global `txn_limbo` used by the tarantool node.

This limitation only applies to the master queue. Once the size of master
queue reaches the maximum value, txn_limbo blocks incoming requests until
some of the transactions in the queue have a quorum of confirmations and
there is free space.

The size limit isn't strict, i.e. if there's at least one free byte, the
whole write request fits and no blocking is involved.

Part of #7486

@TarantoolBot document
Title: new configuration option: 'replication_synchro_queue_max_size'

Since `replication_synchro_timeout` and `IPROTO_RAFT_ROLLBACK` will be
deprecated starting with this patch, synchronous transactions can collect
confirmations indefinitely, the queue can grow very large. You need to be
able to limit the size of a synchronous queue.

`replication_synchro_queue_max_size` puts a limit on the number of
transactions in the master synchronous queue.
`replication_synchro_queue_max_size` is measured in number of bytes to
be written (0 means unlimited, which was the default behaviour before).
This affects only the behavior of the master, and defaults to 16 megabytes.

Now that `replication_synchro_queue_max_size` is set on the master node,
tarantool will discard new transactions that try to queue after the limit
is reached. If a transaction had to be discarded, the user receives an
error message "Synchro queue is full".

The current synchro queue size can be known using
`box.info.synchro.queue.size`:

```lua
tarantool> box.info.synchro
---
- queue:
    owner: 1
    size: 60
    busy: false
    len: 1
    term: 2
  quorum: 2
...

```

NO_CHANGELOG=next commit