txn_limbo: introduce limits on synchro queue
Since `replication_synchro_timeout` and `IPROTO_RAFT_ROLLBACK` will be
deprecated starting with this patch, synchronous transactions can collect
confirmations indefinitely, the queue can grow very large. You need to be
able to limit the size of a synchronous queue.

A new counter is added to `txn_limbo` structure: queue_size (in bytes)
together with a corresponding configuration setting:
`replication_synchro_queue_max_size`. The counter is increased on every
enqueued `txn_limbo_entry`, and decreased once an entry leaves the
`txn_limbo.queue`. Also, a pointer to `journal_entry` has been added to
the `txn_limbo_entry` structure, so that at the time of adding/deleting
an entry to the queue, we have access to the size of the corresponding
entry in the journal.

Actually, the limit is added to an abstract `txn_limbo`, but currently
only works for the global `txn_limbo` used by the tarantool node.

Once the size reaches the maximum value, txn_limbo blocks incoming
requests until some of the transactions in the queue have a quorum of
confirmations and there is free space.

The size limit isn't strict, i.e. if there's at least one free byte, the
whole write request fits and no blocking is involved.

Part of #7486

@TarantoolBot document
Title: new configuration option: 'replication_synchro_queue_max_size'

Since `replication_synchro_timeout` and `IPROTO_RAFT_ROLLBACK` will be
deprecated starting with this patch, synchronous transactions can collect
confirmations indefinitely, the queue can grow very large. You need to be
able to limit the size of a synchronous queue.

`replication_synchro_queue_max_size` puts a limit on the number of
transactions in a synchronous queue.
`replication_synchro_queue_max_size` is measured in number of bytes to
be written (0 means unlimited, which was the default behaviour before).
This affects both the behavior of the master and the behavior of the
replica, and defaults to 16 megabytes.

Now, when `replication_synchro_queue_max_size` is set, tarantool will
ignore new transactions that try to queue once the limit is reached.
If the limit is reached on a replica, as soon as there is free space in
the queue, the replica will process the next incoming transaction from
the replication flow.

NO_CHANGELOG=next commit