limbo: introduce limits on synchro queue
Two new fields added to the structure: the `size` counter and the
`max_size` limit (both in bytes). And also added the corresponding
configuration parameter: `replication_synchro_queue_max_size`.
The counter is increased on every enqueued `txn_limbo_entry`, and
decreased once an entry leaves the `txn_limbo.queue`. Also, the
`approx_len` field has been added to the `txn_limbo_entry` structure,
so that at the time of adding/deleting an entry to the queue, we have
access to the size of the corresponding entry in the journal.

This limitation only applies to the master queue. Once the size of master
queue reaches the maximum value, txn_limbo blocks incoming requests until
some of the transactions in the queue have a quorum of confirmations and
there is free space.

This limitation does not apply during the recovery process, because
otherwise tarantool may fail during the process of the xlog files, if limbo
queue size exceeds `replication_synchro_queue_max_size` and user will have
to pick up the correct value of the `replication_synchro_queue_max_size`
option in order to recover from his xlogs.

The size limit isn't strict, i.e. if there's at least one free byte, the
whole entry fits and no blocking is involved.

Part of #7486

@TarantoolBot document
Title: new configuration option: 'replication_synchro_queue_max_size'
Product: Tarantool
Since: 3.2

`replication_synchro_queue_max_size` puts a limit on the number of
transactions in the master synchronous queue.
`replication_synchro_queue_max_size` is measured in number of bytes to
be written (0 means unlimited, which was the default behaviour before).
This option affects only the behavior of the master, and defaults to
16 megabytes.

Now that `replication_synchro_queue_max_size` is set on the master node,
tarantool will discard new transactions that try to queue after the limit
is reached. If a transaction had to be discarded, user will get an error
message "The synchronous transaction queue is full".

This limitation does not apply during the recovery process.

The current synchro queue size can be known using
`box.info.synchro.queue.size`:

```lua
tarantool> box.info.synchro
---
- queue:
    owner: 1
    size: 60
    busy: false
    len: 1
    term: 2
  quorum: 2
...

```

NO_CHANGELOG=Will be added in another commit