box: forbid TCL statements in transactional triggers
User is not allowed to execute TCL statements in on_commit and
on_rollback triggers - it is documented UB. However, this prohibition
can be easily overseen: one could, for example, try to rollback current
transaction if something is not OK in on_commit trigger. So let's check
if TCL statements are not executed in transactional triggers.

The commit introduces a new helper `txn_check_can_tcl` - it is required
because `txn_check_can_continue` is not suitable for TCL statements: for
example, transaction, aborted due to mvcc conflict, is allowed to call
`box.commit()`, but it will fail with another error (aborted by conflict).

Along the way, the commit removes unnecessary check in function
`box_txn_rollback`.

Closes #7331

NO_DOC=bugfix