Don't re-review backports with doc and infra teams
These changes are anyway approved by these teams before pushing to
`master`.

In rare cases we need to adjust something in a changelog entry or a test
as part of a backport. These changes will require a manual review
request from a backport author.

Kept myself as a reviewer of backports related to the config's code. I
need it to track what is delivered in a feature release and in a bugfix
release.

NO_DOC=no code changes
NO_CHANGELOG=see NO_DOC
NO_TEST=see NO_DOC