test: wait until leaving orphan in cluster:start()
Most of times the instances are already in the `'running'` status after
`cluster:start()`. However, sometimes an instance is not connected yet
to a majority of peers of its replicaset and has the `'orphan'` status.

A test unlikely assumes this situation, because it is rare. Let's wait
for the `'running'` status by default to stay on the safe side.

The new option `wait_until_running` (default is `true`) allows to opt
out from this behavior. It is set to `false` if `wait_until_ready` is
`false`.

This change improves stability of
`config-luatest/anonymous_replica_test.lua`.

Note: `wait_until_ready` waits until `config:info().status` will be
`'ready'` or `'check_warnings'` for all the instances of the cluster.

Next, `wait_until_running` on a replicaset bootstrap waits until all the
instances successfully connect to a majority of peers (see the
`replication.bootstrap_strategy` option for details).

NO_DOC=testing helper change
NO_CHANGELOG=see NO_DOC
NO_TEST=no code is added to tarantool itself