config: introduce config:instance_uri()
This patch introduces the instance_uri() method for the config module.
This method returns the URI of the specified instance from the config.

Needed for #9842

@TarantoolBot document
Title: instance_uri() method for config module

The configuration module's `instance_uri()` method returns the URI of
the instance that is used to create a replicaset or sharding cluster.
This method takes two arguments:
1) `peer` or null to return the URI that is used to create the
   replicaset, or `sharding` to return the URI that is used to create
   the sharding cluster;
2) `options`, which only support `options.instance`, to specify the
   instance whose URI should be returned; by default, the URI of the
   instance on which the function is executed is returned.