trigger: do not return handler from trigger.del
It was mistake to return handler from trigger deletion - if the handler
is not Lua object, its lifetime is independent from Lua, and we
basically return object from its destructor. Let's just return nothing
because checking if a trigger was actually deleted seems to be misusage
of our trigger paradigm.

NO_CHANGELOG=not documented behavior
NO_DOC=bugfix