config: add warnings on skipped URIs
There is a function `find_suitable_uri()` that basically looks for a
URI with non-zero address and port. This patch adds logging that should
ensure that it is easy to understand if a URI was skipped and why.
Note that additional logging is disabled by default and happens only
in sharding and replicaset configuration.

Closes #9644

NO_DOC=internal
NO_TEST=hard to test internal logging

(cherry picked from commit bc18a0547fea0c91651426e1add7558cd8cd6070)