ci: don't send coverage results for submodules
If coverage workflow is triggered via submodule as a part of integration
testing, coverage reports from Tarantool are sent to coveralls as
reports from the submodule. Our fork of LuaJIT as well as Tarantool uses
Coveralls, this will avoid reporting Tarantool's coverage to LuaJIT's
namespace in Coveralls.

To avoid this misbehaviour, run the "upload report" step only when there
is no submodule to bump (specified as `inputs.submodule` [1]) as a part
of the reusable workflow.

[1]: https://docs.github.com/en/actions/creating-actions/metadata-syntax-for-github-actions#example-specifying-inputs

NO_DOC=CI
NO_TEST=CI
NO_CHANGELOG=CI