func_adapter_lua: allow to pass arguments on Lua stack from fiber
Currently, passing arguments on Lua stack from fiber to func_adapter_lua
will lead to problems, so let's manually check if `args` is `port_lua`,
and, if that's so, check if Lua-stack is not the same as stored in fiber.
If the check fails, we will create a new one. This logic allows to call
func_adapter_lua without thinking about sharing the same Lua stack.

We could create the new Lua-stack on every func_adapter call, but it's
costly. For example, now one empty on_replace trigger slows Tarantool
down by 27%, but without the optimization slowdown is about 42%.

NO_CHANGELOG=internal
NO_DOC=internal