bootstrap: disable dd checks on box.internal.bootstrap
Lately we restricted access to system spaces for users. So, in order to
allow schema upgrade to modify them, data dictionary checks are disabled
when upgrade is started for the current fiber and enabled back when it is
completed. But we forgot to do the same with `box.internal.bootstrap`, so
now it fails on access check. The commit fixes this mistake.

NO_TEST=not tested tool
NO_CHANGELOG=internal
NO_DOC=internal