main: add --failover into CLI options help message
The failover coordinator is a service provided by Tarantool Enterprise
Edition. It is going to be released in the 3.1.0 version. Let's add the
`--failover` option into the help message and the man page.

This reverts commit af4bad4364f1 ("main: hide --failover CLI option").

Part of tarantool/tarantool-ee#564

NO_DOC=the documentation request for the failover coordinator at whole
       will be added manually
NO_CHANGELOG=this is EE feature, no reason to highlight it in CE release
             notes