xrow: export vclock serialization functions to header
When implementing persistent gc consumers, we will need to serialize
vclocks. We already have such helpers, but they are private for
`xrow.c`. So the commit simply puts their declarations to the
appropriate header and marks them as non-static.

Also, the commit adds a check for correct lsn and replica id - it is
needed to safely decode user tuples.

Part of #10154

NO_CHANGELOG=internal
NO_DOC=internal