ci: add timeouts for workflow jobs
By default, each job in a workflow can run for up to 6 hours of the
execution time. If a job reaches this limit, the job is terminated by
GitHub automatically and fails to complete. This patch sets job timeouts
to 60 minutes to avoid waiting for jobs to complete for 6 hours.

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci