test: fix flaky persistent triggers test
The replication test of persistent triggers was waiting only for
the persistent triggers to arrive on replica, so the replica tried
to write to the space which was not created there yet. Let's wait
for all changes to arrive to make the test stable.

Closes #9967

NO_CHANGELOG=test
NO_DOC=test

(cherry picked from commit 927e351625292c20ff6d07cc70164725fce8b338)