box: deprecate wal_cleanup_delay
The commit deprecates option wal_cleanup_delay - it was introduced as a
workaround for a problem, which is solved by persistent gc state.

However, we automatically set wal_cleanup_delay to non-zero value if
current schema does not support _gc_consumers space and the instance
is a part of a replicaset. It prevents xlogs from being cleaned up on
upgrade to version with deprecated but still supported
wal_cleanup_delay option.

NO_DOC=tmp