box: deprecate wal_cleanup_delay
The commit deprecates option wal_cleanup_delay - it was introduced as a
workaround for a problem, which is solved by persistent gc state.

Since wal_cleanup_delay doesn't actually work when all replicas have gc
consumers (any standard situation), let's simply set old default value
when initializing gc subsystem - non-zero value is needed to retain xlogs
during schema upgrade. Then, we can simply log a warning or raise an error
depending on new compat option `wal_cleanup_delay_deprecation` on any
attempt to set this option. Default option for this value is `nil` in order
not to trigger option setter when it's not set manually.

What's about config part, we set no default value there because even
setting `box.NULL` would trigger option setter on reload which is bad
because user would receive a warning or an error even without touching
the option.

Follow-up #10154

@TarantoolBot document
Title: Deprecate wal_cleanup_delay

The option `wal_cleanup_delay` is deprecated starting from Tarantool 3.2.0.
The new compat module option `wal_cleanup_delay_deprecation` was
introduced to control whether the option is still available.
With the old behavior, which is the default in Tarantool 3.2.0,
the option is still available and has the same default value, but a
deprecation warning is logged when it is set manually. With the new
behavior, an error is raised on any attempt to set this option, but its
default value isn't actually changed.

(Please create https://tarantool.io/compat/wal_cleanup_delay_deprecation)

We are planning to switch the compat option to the new behavior
starting from Tarantool 4.0 with the ability to revert to the
old behavior. Starting from Tarantool 5.0 we are planning to
drop `wal_cleanup_delay` completely - the behavior will be the same
as when this option is set to zero.